Add some opcodes fixes for 5.4.0 17399: - CMSG_ITEM_REFUND - CMSG_ITEM_REFUND_INFO - SMSG_ITEM_REFUND_INFO_RESPONSE - CMSG_KEEP_ALIVE - CMSG_UNREGISTER_ALL_ADDON_PREFIXES - SMSG_ITEM_REFUND_RESULT - SMSG_UPDATE_CURRENCY - SMSG_UPDATE_CURRENCY_WEEK_LIMIT - SMSG_SELL_ITEM - Fix Week cap of Valor Points - CMSG_PAGE_TEXT_QUERY - SMSG_PAGE_TEXT_QUERY_RESPONSE - Fix mail money is not deleted if gold cap is reached - Fix max bank money is up to 999.999.99.99 - SMSG_GUILD_XP - Switch Guild loading in PreparedStatements, fix money loading - Refresh guild roster after member kick - Implement SMSG_GUILD_SEND_MEMBER_LEAVE

+++ src/server/game/Scripting/ScriptMgr.cpp
@@ -1393,7 +1393,7 @@ void ScriptMgr::OnGuildMemberWitdrawMoney(Guild* guild, Player* player, uint64 &
     FOREACH_SCRIPT(GuildScript)->OnMemberWitdrawMoney(guild, player, amount, isRepair);
 }
 
-void ScriptMgr::OnGuildMemberDepositMoney(Guild* guild, Player* player, uint32 &amount)
+void ScriptMgr::OnGuildMemberDepositMoney(Guild* guild, Player* player, uint64 &amount)
 {
     FOREACH_SCRIPT(GuildScript)->OnMemberDepositMoney(guild, player, amount);
 }
